+++ tests/testthat/test-validate_config.R
@@ -1,4 +1,5 @@
 test_that("Config validated successfully", {
+  skip_if_offline()
   expect_true(suppressMessages(validate_config(
     hub_path = system.file(
       "testhubs/simple/",
@@ -25,6 +26,7 @@ test_that("Config validated successfully", {
 
 
 test_that("Config for samples handled succesfully", {
+  skip_if_offline()
   config_path <- testthat::test_path("testdata", "tasks-samples-pass.json")
   out <- suppressMessages(validate_config(
     config_path = config_path,
@@ -34,6 +36,7 @@ test_that("Config for samples handled succesfully", {
   expect_true(out)
 })
 test_that("Missing files returns an invalid config with an immediate message", {
+  skip_if_offline()
   tmp <- withr::local_tempdir()
   suppressMessages({
     expect_message(out <- validate_config(hub_path = tmp), "File does not exist")
@@ -41,6 +44,7 @@ test_that("Missing files returns an invalid config with an immediate message", {
   expect_false(out)
 })
 test_that("Config for samples fail correctly", {
+  skip_if_offline()
   config_path <- testthat::test_path("testdata", "tasks-samples-error-range.json")
   out <- suppressWarnings(validate_config(
     config_path = config_path,
@@ -63,6 +67,7 @@ test_that("Config for samples fail correctly", {
 
 
 test_that("Config errors detected successfully", {
+  skip_if_offline()
   config_path <- testthat::test_path("testdata", "tasks-errors.json")
   out <- suppressWarnings(validate_config(config_path = config_path))
   expect_snapshot(out)
@@ -72,6 +77,7 @@ test_that("Config errors detected successfully", {
 
 
 test_that("Dynamic config errors detected successfully by custom R validation", {
+  skip_if_offline()
   config_path <- testthat::test_path("testdata", "tasks-errors-rval.json")
   out <- suppressWarnings(validate_config(config_path = config_path))
   expect_snapshot(out)
@@ -80,6 +86,7 @@ test_that("Dynamic config errors detected successfully by custom R validation", 
 })
 
 test_that("Reserved hub variable task id name detected correctly", {
+  skip_if_offline()
   config_path <- testthat::test_path("testdata", "tasks-errors-rval-reserved.json")
   out <- suppressWarnings(validate_config(config_path = config_path))
   expect_snapshot(out)
@@ -88,6 +95,7 @@ test_that("Reserved hub variable task id name detected correctly", {
 })
 
 test_that("NULL target keys validated successfully", {
+  skip_if_offline()
   config_path <- testthat::test_path("testdata", "tasks_null_rval.json")
   out <- suppressMessages(validate_config(config_path = config_path))
   expect_true(out)
@@ -95,12 +103,14 @@ test_that("NULL target keys validated successfully", {
 
 
 test_that("Bad schema_version URL errors successfully", {
+  skip_if_offline()
   config_path <- testthat::test_path("testdata", "schema_version-errors.json")
   expect_error(validate_config(config_path = config_path))
 })
 
 
 test_that("Additional properties error successfully", {
+  skip_if_offline()
   config_path <- testthat::test_path("testdata", "tasks-addprop.json")
   out <- suppressWarnings(validate_config(config_path = config_path))
   expect_snapshot(out)
@@ -110,6 +120,7 @@ test_that("Additional properties error successfully", {
 
 
 test_that("Duplicate values in individual array error successfully", {
+  skip_if_offline()
   config_path <- testthat::test_path("testdata", "dup-in-array.json")
   out <- suppressWarnings(validate_config(config_path = config_path))
   expect_snapshot(out)
@@ -118,6 +129,7 @@ test_that("Duplicate values in individual array error successfully", {
 })
 
 test_that("Duplicate values across property error successfully", {
+  skip_if_offline()
   config_path <- testthat::test_path("testdata", "dup-in-property.json")
   out <- suppressWarnings(validate_config(config_path = config_path))
   expect_snapshot(out)
@@ -126,6 +138,7 @@ test_that("Duplicate values across property error successfully", {
 })
 
 test_that("Inconsistent round ID variables across model tasks error successfully", {
+  skip_if_offline()
   config_path <- testthat::test_path("testdata", "round-id-inconsistent.json")
   out <- suppressWarnings(validate_config(config_path = config_path))
   expect_snapshot(out)
@@ -141,6 +154,7 @@ test_that("Inconsistent round ID variables across model tasks error successfully
 
 
 test_that("Duplicate round ID values across rounds error successfully", {
+  skip_if_offline()
   config_path <- testthat::test_path("testdata", "dup-in-round-id.json")
   out <- suppressWarnings(validate_config(config_path = config_path))
   expect_snapshot(out)
@@ -149,6 +163,7 @@ test_that("Duplicate round ID values across rounds error successfully", {
 })
 
 test_that("All null task IDs error successfully", {
+  skip_if_offline()
   config_path <- testthat::test_path("testdata", "both_null_tasks_all.json")
   out <- suppressWarnings(validate_config(config_path = config_path))
   expect_snapshot(out)
@@ -157,6 +172,7 @@ test_that("All null task IDs error successfully", {
 })
 
 test_that("Old orgname config validates successfully", {
+  skip_if_offline()
   config_path <- testthat::test_path("testdata", "task-old-orgname.json")
   out <- suppressMessages(validate_config(config_path = config_path))
   expect_snapshot(out)
@@ -165,6 +181,7 @@ test_that("Old orgname config validates successfully", {
 })
 
 test_that("target keys error if arrays passed", {
+  skip_if_offline()
   config_path <- testthat::test_path(
     "testdata",
     "tasks-target-key-array-v4.json"
@@ -188,4 +205,45 @@ test_that("target keys error if arrays passed", {
   # but not about target keys.
   expect_snapshot(attr(out_v3, "errors"))
   expect_false(out_v3)
+})
+
+test_that("v4 validation works", {
+  skip_if_offline()
+  config_path <- testthat::test_path(
+    "testdata",
+    "v4-tasks.json"
+  )
+  expect_true(
+    suppressMessages(
+      validate_config(
+        config_path = config_path,
+        branch = "br-v4.0.0"
+      )
+    )
+  )
+  config_path <- testthat::test_path("testdata", "v4-tasks-fail.json")
+  expect_false(
+    suppressMessages(
+      v4_fail <- validate_config(config_path = config_path, branch = "br-v4.0.0")
+    )
+  )
+  expect_snapshot(extract_error_tbl_cols(v4_fail))
+
+  # Ensure dynamic validation works and catches invalid derived task IDs at config
+  # and round level.
+  config_path <- testthat::test_path("testdata", "v4-tasks-fail-dynamic.json")
+  expect_false(
+    suppressMessages(
+      v4_fail_dynamic <- validate_config(config_path = config_path, branch = "br-v4.0.0")
+    )
+  )
+  expect_snapshot(
+    extract_error_tbl_cols(v4_fail_dynamic, c(
+      "message", "data"
+    ))
+  )
+  expect_equal(
+    extract_error_tbl_cols(v4_fail_dynamic, c("instancePath"))$instancePath,
+    structure(c("/rounds/0/derived_task_ids", "/derived_task_ids"), class = c("glue", "character"))
+  )
 })